requirements.txt
streamlit==1.29.0
Pillow==10.1.0
firebase-admin==6.3.0
google-cloud-vision==3.4.5
anthropic==0.8.1
python-dotenv==1.0.0